requirements.txt
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.1.0
requests==2.31.0
